+++ tests/frost.rs
@@ -10,7 +10,7 @@ fn check_sign_with_dealer() {
     let threshold = 3;
     let (shares, pubkeys) = frost::keygen_with_dealer(numsigners, threshold, &mut rng).unwrap();
 
-    let mut nonces: HashMap<u32, Vec<frost::SigningNonces>> =
+    let mut nonces: HashMap<u8, Vec<frost::SigningNonces>> =
         HashMap::with_capacity(threshold as usize);
     let mut commitments: Vec<frost::SigningCommitments> = Vec::with_capacity(threshold as usize);
